Fix board state rotations.
