+++ neuralknight/tests/test_board.py
@@ -295,18 +295,18 @@ def test_board_lookahead_player_is_constant(start_board):
         (0, 9, 9, 9, 9, 9, 9, 9),
         (13, 7, 3, 11, 5, 3, 7, 13))
     assert states[1].board == (
-        (12, 0, 2, 10, 4, 2, 6, 12),
-        (8, 8, 8, 8, 8, 8, 8, 8),
-        (6, 0, 0, 0, 0, 0, 0, 0),
+        (12, 6, 2, 10, 4, 2, 6, 12),
+        (8, 8, 8, 8, 8, 8, 8, 0),
+        (0, 0, 0, 0, 0, 0, 0, 8),
         (0, 0, 0, 0, 0, 0, 0, 0),
         (0, 0, 0, 0, 0, 0, 0, 0),
         (9, 0, 0, 0, 0, 0, 0, 0),
         (0, 9, 9, 9, 9, 9, 9, 9),
         (13, 7, 3, 11, 5, 3, 7, 13))
     assert states[2].board == (
-        (12, 0, 2, 10, 4, 2, 6, 12),
-        (8, 8, 8, 8, 8, 8, 8, 8),
-        (6, 0, 0, 0, 0, 0, 0, 0),
+        (12, 6, 2, 10, 4, 2, 6, 12),
+        (8, 8, 8, 8, 8, 8, 8, 0),
+        (0, 0, 0, 0, 0, 0, 0, 8),
         (0, 0, 0, 0, 0, 0, 0, 0),
         (9, 0, 0, 0, 0, 0, 0, 0),
         (0, 0, 0, 0, 0, 0, 0, 0),
