+++ neuralknight/tests/test_mock_board_integration.py
@@ -81,10 +81,11 @@ def test_agent_play_through(testapp):
         (0, 0, 9, 0, 0, 0, 0, 0),
         (9, 9, 0, 9, 9, 9, 9, 9),
         (13, 7, 3, 5, 11, 3, 7, 13))]
-    player2_moves = [player2_moves[0]]
+    player2_moves = []
     game = testapp.post_json('/v1.0/games').json
     player1 = MockAgent(testapp, player1_moves, game['id'], 1)
     player2 = MockAgent(testapp, player2_moves, game['id'], 2)
-    assert len(player1.args) == 2
-    assert len(player2.args) == 0
-    assert not player2.past_end
+    assert len(player1.args) == 1
+    assert len(player2.args) == 1
+    assert not player1.past_end
+    assert player2.past_end
